WIP refactor: move restriction impl to new file
